Fix warnings from Open Watcom.
git-svn-id: https://svn.code.sf.net/p/freedos/svn/freecom/trunk@1586 6ac86273-5f31-0410-b378-82cca8765d1b

+++ include/context.h
@@ -73,7 +73,7 @@ extern ctxt_info_t ctxt_info[];
 	because just <tag>= is not valid as _item_, but may contain
 	status information */
 #define ctxtProbeItemTag(segm,ofs,tag)							\
-	(peekb((segm), (ofs)) == (unsigned char)(Context_Tag)(tag)	\
+	(peekb((segm), (ofs)) == (unsigned char)(Context_Tag)(int)(tag)	\
 		 && peekb((segm), (ofs) + 1) != '=')
 #define CTXT_ITEMNAME_LENGTH (sizeof(unsigned) * 2 + 2)
